+++ console/module/device/src/main/java/org/eclipse/kapua/app/console/module/device/client/device/DeviceGrid.java
@@ -231,9 +231,14 @@ public class DeviceGrid extends EntityGrid<GwtDevice> {
     @Override
     protected void selectionChangedEvent(GwtDevice selectedItem) {
         super.selectionChangedEvent(selectedItem);
-        getToolbar().getEditEntityButton().setEnabled(currentSession.hasPermission(DeviceSessionPermission.write()));
-        getToolbar().getAddEntityButton().setEnabled(currentSession.hasPermission(DeviceSessionPermission.write()));
-        getToolbar().getDeleteEntityButton().setEnabled(currentSession.hasPermission(DeviceSessionPermission.delete()));
+        if (selectedItem != null) {
+            getToolbar().getEditEntityButton().setEnabled(currentSession.hasPermission(DeviceSessionPermission.write()));
+            getToolbar().getAddEntityButton().setEnabled(currentSession.hasPermission(DeviceSessionPermission.write()));
+            getToolbar().getDeleteEntityButton().setEnabled(currentSession.hasPermission(DeviceSessionPermission.delete()));
+        } else {
+            getToolbar().getEditEntityButton().setEnabled(false);
+            getToolbar().getDeleteEntityButton().setEnabled(false);
+        }
     }
 
     @Override
